Add class FMPathPermission & Tests
I added the class FMPathPermission and corresponding basic tests.
The FMPathPermission keeps a list of Paths and a list of permission to
apply to those paths.
It also holds the information (ACRule) how to deal with inherited
permissions.
Besides that I put the ClassDefinitions.ps1 into the \public folder
for it offers public functions.