Add class FMDirectory and helper function
I added a class FMDirectory to hold a complete directory structure
with a root path and some derived child path(s).
The root path is configured for full inheritance, whild the children
are configured to break in inheritance and transfer permissions in
explicit ones.
There is also a helper function New-FMDirectory to create such an
object with ease.
Surely some tests are added.