Add test for FMPathPermission class and Should -ContainACE
I added some unit tests for the FMPathPermission class and wrote a
helper assertion, that checks, if the given permission is contained in
the acl (checking identity, permission and inheritance). The permission
check is not precise, because I'm checking for my permission, i.e.
'Read', while the actual permission in the filesystem is 'Read,
Synchonize', but it's sufficient for my purposes.